Fix project (added user access rights check) - changed type fields of id (to long); - added field User user to Card (added to scripts) - added user access rights check; - changed implementation of methods deleteCardById and leaveByCardId; - changed tests

+++ src/main/java/com/greetingcard/web/servlet/card/GetCardsServlet.java
@@ -20,12 +20,12 @@ public class GetCardsServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
         User user = (User) request.getSession().getAttribute("user");
-        int userId = user.getId();
+        long userId = user.getId();
 
         Map<Card, Role> cards = cardService.getCards(userId, request.getParameter("cards-type"));
 
         Map<String, Object> parameters = new HashMap<>();
-        parameters.put("cards", cards);
-        PageGenerator.getInstance().process("after-login", parameters, request, response);
+        parameters.put("cards" , cards);
+        PageGenerator.getInstance().process("after-login" , parameters, request, response);
     }
 }
